requirements.txt
APScheduler==3.10.4
Flask-APScheduler==1.13.1
Flask-Cors==5.0.0
MarkupSafe==3.0.2
PySocks==1.7.1
SQLAlchemy==2.0.36
argparse==1.4.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cors==1.0.1
filelock==3.16.1
Flask==3.0.3
future==1.0.0
gevent==24.10.3
greenlet==3.1.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
mysql-connector-python==9.1.0
pip==24.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
schedule==1.2.2
setuptools==75.2.0
six==1.16.0
soupsieve==2.6
tldextract==5.1.2
typing-extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.0
